requirements.txt
--only-binary matplotlib
--only-binary numpy
--only-binary pandas

affine==2.3.0
attrs==19.3.0
backcall==0.1.0
bleach==3.1.4
branca==0.3.1
Cartopy==0.17.0
click==7.1.1
click-plugins==1.1.1
cligj==0.5.0
contextily==0.99.0
cycler==0.10.0
Cython==0.29.16
decorator==4.4.2
defusedxml==0.6.0
Deprecated==1.2.7
descartes==1.1.0
entrypoints==0.3
Fiona==1.8.13.post1
geographiclib==1.50
geopandas==0.7.0
geopy==1.21.0
importlib-metadata==1.6.0
ipykernel==5.2.0
ipyleaflet==0.12.3
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.2
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
mapclassify==2.2.0
MarkupSafe==1.1.1
matplotlib==3.2.1
mercantile==1.1.2
mistune==0.8.4
more-itertools==8.2.0
munch==2.5.0
nbconvert==5.6.1
nbformat==5.0.5
notebook==6.0.3
numpy==1.18.2
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.1
plotly==4.6.0
pluggy==0.13.1
prometheus-client==0.7.1
prompt-toolkit==3.0.5
ptyprocess==0.6.0
py==1.8.1
Pygments==2.6.1
pyparsing==2.4.6
pyproj==2.6.0
pyrsistent==0.16.0
pyshp==2.1.0
pytest==5.4.1
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
qtconsole==4.7.2
QtPy==1.9.0
rasterio==1.1.3
retrying==1.3.3
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
Shapely==1.7.0
six==1.14.0
snuggs==1.4.7
streamlit
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
traittypes==0.2.1
wcwidth==0.1.9
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.12.1
xarray==0.15.1
zipp==3.1.0
epimodels>=0.3.2

